requirements.txt
# FastAPI и зависимости
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.5.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Утилиты
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Тестирование
pytest==8.0.0
httpx==0.26.0

# Форматирование и линтинг
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Шаблоны
jinja2==3.1.2 